fix: opentelemetry broken traces
Activity.Current is a static variable that uses AsyncLocal internally,
which means that it flows into async calls,
but not back out into the caller (message consumer and producer).

As so, the activity.current is null after the produce and consume of the
message, which means that all the spans created after it will be
generate a new trace because the context is not being propagated.