fix: opentelemetry broken traces
Activity.Current is a static variable that uses AsyncLocal internally,
which means that it flows into children async calls, but not back to the
caller (message consumer and producer).
With the previous implementation, the Activity.Current is null after the
produce and consume of messages, which means that all spans created
after it will generate new trace information because the context is not
being propagated.

This change fixes the problem firing sync events and making the
OpenTelemetry handlers subscribe them.